Moved functions into another file, removed fields from agents
